+++ src/main/kotlin/io/acari/doki/actions/MaterialFilesconsAction.kt
@@ -2,14 +2,13 @@ package io.acari.doki.actions
 
 import com.intellij.openapi.actionSystem.AnActionEvent
 import io.acari.doki.config.ThemeConfig
-import io.acari.doki.settings.actors.StickerActor
-import io.acari.doki.stickers.CurrentSticker
+import io.acari.doki.settings.actors.MaterialIconsActor
 
-class SwapStickerAction : BaseToggleAction() {
+class MaterialFilesconsAction : BaseToggleAction() {
   override fun isSelected(e: AnActionEvent): Boolean =
-    ThemeConfig.instance.currentSticker == CurrentSticker.SECONDARY
+    ThemeConfig.instance.isMaterialFiles
 
   override fun setSelected(e: AnActionEvent, state: Boolean) {
-      StickerActor.swapStickers(state)
+      MaterialIconsActor.enableFileIcons(state)
   }
 }