+++ src/main/kotlin/io/acari/doki/notification/UpdateNotification.kt
@@ -14,13 +14,14 @@ val UPDATE_MESSAGE: String = """
             </ul>
         </li>
         <li>Android Studio 4.0 Support</li>
+        <li>2020 EAP Support</li>
       </ul>
       <br>Please see the <a href="https://github.com/cyclic-reference/ddlc-jetbrains-theme/blob/master/changelog/CHANGELOG.md">Changelog</a> for more details.
       <br>
       Thanks again for downloading <b>The Doki Theme</b>! •‿•<br>
 """.trimIndent()
 
-const val CURRENT_VERSION = "6.0.0"
+const val CURRENT_VERSION = "6.0.1"
 
 object UpdateNotification {