+++ changelog/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 ---
 
+# 6.0.1
+
+- 2020 EAP Build Support
+    - Settings menu is currently unavailable in the 2020 builds, all actions are in the tool menu theugh.
+- Fixed issue with not being able to select a Doki Theme from the settings menu.
+
 # 6.0.0
 
 - Added new themes based on characters from various anime series!